記事データの型定義場所変更
