+++ src/app/plugins/HtmlToolbarPlugin.tsx
@@ -7,20 +7,7 @@ import { $generateNodesFromDOM } from '@lexical/html';
 import { CiExport, CiImport } from "react-icons/ci";
 import styles from "./CommonToolbar.module.scss"
 
-const categories: { [key: string]: string } = {
-  newcomer: "新入生の方へ",
-  oriter: "オリターの方へ",
-  circle: "サークルの方へ",
-  committee: "当委員会について",
-};
-export type articleType = {
-  title: string; // 記事タイトル (pathにもなる)
-  category: (typeof categories)[number]; // 大枠のカテゴリ
-  subCategory: string; // 各ページ内で見出しでまとめる用
-  date: string; // 更新日
-  article: string; // html形式
-  shown: boolean; // 公開するかどうか
-}
+import type { ArticleType } from "../types";
 
 // Export
 export const Export = (editor: any, exporterAsHTML?: Function) => {
@@ -45,7 +32,7 @@ export const Import = (editor: any, defaultContentAsHTML?: string) => {
 };
 
 //HTMLToolbarPlugin
-export const HTMLToolbarPlugin: FC<{articleData: articleType | undefined}> = (props) => {
+export const HTMLToolbarPlugin: FC<{articleData: ArticleType | undefined}> = (props) => {
   const EXPORT_COMMAND : LexicalCommand<Function> = createCommand();
   const IMPORT_COMMAND: LexicalCommand<string> = createCommand();
   const [ editor ] = useLexicalComposerContext();
